// types/generated/ContentQueryApiV2.ts
export type DistributionDate = string;
export type DistributionStatus = "READY" | "NOT_READY" | "SENT" | "UNPUBLISHED";